% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_compare.R
\name{graph_compare}
\alias{graph_compare}
\title{Compare distributions across selected groups}
\usage{
graph_compare(data, groupColumn, input_var, input_mod, input_pct = "percent")
}
\arguments{
\item{data}{Report-level data.}

\item{groupColumn}{Column containing the categories to display.}

\item{input_var}{Column defining the groups being compared.}

\item{input_mod}{Selected group values.}

\item{input_pct}{Counts (niv or nb) or percentages within each group.}
}
\value{
A Highcharter widget.
}
\description{
Compare distributions across selected groups
}
