% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_panel_exploration.R
\name{nav_panel_exploration}
\alias{nav_panel_exploration}
\title{Analysis views}
\usage{
nav_panel_exploration()
}
\value{
A bslib tab set containing overview, territory, comparison and data views.
}
\description{
Analysis views
}
