% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maj_data.R
\name{maj_data}
\alias{maj_data}
\title{Title}
\usage{
maj_data(
  url = "https://www.data.gouv.fr/fr/datasets/r/106add51-e925-4121-b0e4-81489c26a43f",
  destination = "signalconso.csv"
)
}
\description{
Title
}
